unify naming rules of all File
